created errorHandlingMiddleware

+++ services/classes/UserDeviceService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Identity.Client;
+using SmartHomeAsistent.CustomExceptions;
 using SmartHomeAsistent.DTO;
 using SmartHomeAsistent.Entities;
 using SmartHomeAsistent.services.interfaces;
@@ -18,12 +19,13 @@ namespace SmartHomeAsistent.services.classes
 
         public async Task<bool> AddUserDevice(UserDeviceDTO userDeviceDTO)
         {
+            
             if(_context.UserDevices.Any(x=>x.UserId == userDeviceDTO.UserId && x.DeviceId == userDeviceDTO.DeviceId))
-                throw new Exception("Такое устройство уже привязано к пользователю");
+                throw new BadRequestException("Такое устройство уже привязано к пользователю");
             User user =await _context.Users.FirstOrDefaultAsync(x => x.Id == userDeviceDTO.UserId && x.Hidden == false && x.IsBlocked == false)
-                ?? throw new Exception("Пользователь не найден");
+                ?? throw new NotFoundException("Пользователь не найден");
             Device device =await _context.Devices.FirstOrDefaultAsync(x => x.Id == userDeviceDTO.DeviceId )
-                ?? throw new Exception("Устройство не найдено");
+                ?? throw new NotFoundException("Устройство не найдено");
 
             UserDevice userDevice = new()
             {
@@ -39,8 +41,13 @@ namespace SmartHomeAsistent.services.classes
 
         public async Task<bool> DeleteUserDevice(int userId, int deviceId)
         {
+            if (userId <= 0)
+                throw new ValidationException("Невалидный Id пользователя");
+            if (deviceId <= 0)
+                throw new ValidationException("Невалидный Id устройства");
+
             UserDevice userDevice = await _context.UserDevices.FirstOrDefaultAsync(x => x.UserId == userId && x.DeviceId == deviceId)
-                ?? throw new Exception("Сущность не найдена");
+                ?? throw new NotFoundException("Сущность не найдена");
 
             _context.UserDevices.Remove(userDevice);
             await _context.SaveChangesAsync();
@@ -50,20 +57,29 @@ namespace SmartHomeAsistent.services.classes
 
         public async Task<UserDevice> GetUserDeviceById(int userId, int deviceId)
         {
+            if (userId <= 0)
+                throw new ValidationException("Невалидный Id пользователя");
+            if (deviceId <= 0)
+                throw new ValidationException("Невалидный Id устройства");
+
             UserDevice userDevice = await _context.UserDevices.FirstOrDefaultAsync(x => x.UserId == userId && x.DeviceId == deviceId)
-                ?? throw new Exception("Сущность не найдена");
+                ?? throw new NotFoundException("Сущность не найдена");
 
             return userDevice;
         }
 
         public async Task<List<UserDevice>> GetUserDevicesByDeviceId(int deviceId)
         {
+            if (deviceId<=0)
+                throw new ValidationException("Невалидный Id устройства");
             var userDevices =await _context.UserDevices.Where(x => x.DeviceId == deviceId).ToListAsync();
             return userDevices;
         }
 
         public async Task<List<UserDevice>> GetUserDevicesByUserId(int userId)
         {
+            if (userId <= 0)
+                throw new ValidationException("Невалидный Id пользователя");
             var userDevices = await _context.UserDevices.Where(x => x.UserId == userId).ToListAsync();
             return userDevices;
         }
@@ -71,7 +87,7 @@ namespace SmartHomeAsistent.services.classes
         public async Task<bool> UpdateUserDevice(int userId, int deviceId, UserDeviceDTO userDeviceDTO)
         {
             UserDevice userDevice = await _context.UserDevices.FirstOrDefaultAsync(x => x.UserId == userId && x.DeviceId == deviceId)
-               ?? throw new Exception("Сущность не найдена");
+               ?? throw new NotFoundException("Сущность не найдена");
 
             userDevice.UserId = userDeviceDTO.UserId;
             userDevice.DeviceId = userDeviceDTO.DeviceId;
